[BEAM-3129] Restored duplicate button feature
